+++ scripts/util.py
@@ -17,7 +17,7 @@ import time
 AssetDesc = namedtuple('AssetDesc', 'region asset_id project')
 
 #dxda_version = "v0.2.2"
-dxda_version = "20190830190211_55ea371"
+dxda_version = "20190903204842_c71ebee"
 max_num_retries = 5
 
 def dxWDL_jar_path(top_dir):
@@ -128,7 +128,7 @@ def _download_dxda_into_resources(top_dir):
             trg_dxda_tar])
     else:
         # A snapshot of the download-agent development branch
-        command = """sudo  docker run --entrypoint=\'\' dnanexus/dxda:{} cat /builds/dx-download-agent-linux.tar > {}""".format(dxda_version, trg_dxda_tar)
+        command = """sudo  docker run --rm --entrypoint=\'\' dnanexus/dxda:{} cat /builds/dx-download-agent-linux.tar > {}""".format(dxda_version, trg_dxda_tar)
         p = subprocess.Popen(command, universal_newlines=True, shell=True,
                              stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
         text = p.stdout.read()