+++ src/main/scala/dxWDL/compiler/Native.scala
@@ -543,6 +543,12 @@ case class Native(dxWDLrtId: Option[String],
             "release" -> JsString(Utils.UBUNTU_VERSION),
         )
 
+        // Add default timeout
+        val defaultTimeout : Map[String, JsValue] =
+            DxRunSpec(None, None, None, Some(DxTimeout(Some(Utils.DEFAULT_APPLET_TIMEOUT_IN_DAYS),
+                                                       Some(0),
+                                                       Some(0)))).toRunSpecJson
+
         // Start with the default dx-attribute section, and override
         // any field that is specified in the individual task section.
         val extraRunSpec : Map[String, JsValue] = extras match {
@@ -567,7 +573,7 @@ case class Native(dxWDLrtId: Option[String],
                 Map.empty
             }
 
-        val runSpecWithExtras = runSpec ++ extraRunSpec ++ taskSpecificRunSpec
+        val runSpecWithExtras = runSpec ++ defaultTimeout ++ extraRunSpec ++ taskSpecificRunSpec
 
         // - If the docker image is a tarball, add a link in the details field.
         val dockerFile: Option[DXFile] = applet.docker match {