+++ RELEASE_NOTES.md
@@ -1,8 +1,8 @@
 # Release Notes
 
 ## 1.20b
-- Another experimental version of dx-download-agent
-- By default, setting a 48 hour timeout for applets. This can be modified as needed by setting a different timeout in the [extras file](./doc/ExpertOptions.md#Setting-dnanexus-specific-attributes-for-tasks).
+- The default timeout limit for tasks is 48 hours. It can be overriden by setting a different timeout in the [extras file](./doc/ExpertOptions.md#Setting-dnanexus-specific-attributes-for-tasks).
+- Fixing a bug where the database is locked, in the [dx-download-agent](https://github.com/dnanexus/dxda/).
 
 ## 1.20a
 - An experimental version of dx-download-agent